GV-623 added password change validators and typescript in the PasswordChangeForm
